Set low watermark of prekeys in own bundle to 90%
The old value of 25 (of 100 keys total) means, that prior to
regenerating new keys, there will be a 75% chance that a sender picks an
unusable key. Normally that triggers a keytransport element from the
receiver, repairing the broken session / creating a new session, but if
the receiver is offline, that won't happen and all messages sent until
the receiver comes online, will be lost.